新的接口匹配规则

+++ app/src/main/java/com/chinamobile/wanapp/baen/TaskData.java
@@ -99,7 +99,7 @@ public class TaskData implements Serializable {
         return jobTags;
     }
 
-    public int getJzGain() {
+    public double getJzGain() {
         return jzGain;
     }
